restructure match and range filters
